feat: add `get_segmented_dti_abnormal_values` to `MRIExam` class
Needed to relocate DTI lesion entities to mri module in order to avoid circular imports:
- Migrate `DTIAbnormalValues`, `AbnormalValueType`, and `AbnormalVoxelData` from `dti_lesions.py` to `mri.py`.
- Update all module imports and references to reflect the new structure.
- Remove unused `dti_lesions.py` file.
